third_party/WebKit/LayoutTests/external/wpt/service-workers/service-worker/fetch-event-reserved-client.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>FetchEvent: reserved client</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/get-host-info.sub.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<body>
<script>
promise_test(t => {
    const kScope = 'resources/';
    const kScript = 'resources/fetch-event-reserved-client-worker.js';
    let data;
    let worker;
    return service_worker_unregister_and_register(t, kScript, kScope)
      .then(registration => {
          promise_test(() => {
              return registration.unregister();
            }, 'restore global state');
          worker = registration.installing;
          return wait_for_state(t, worker, 'activated');
        })
      .then(() => {
          return with_iframe(kScope + 'iframe');
        })
      .then(frame => {
          add_completion_callback(() => { frame.remove(); });
          data = JSON.parse(frame.contentDocument.body.textContent);
          const saw_message = new Promise(resolve => {
              navigator.serviceWorker.addEventListener('message', e => {
                  resolve(e.data);
                });
            });
          worker.postMessage(frame.contentWindow.location.href);
          return saw_message;
        })
      .then(message => {
          assert_equals(message.clientId, data.reservedClientId,
              'reserved client id should match the resulting document');
        })
  });
</script>
<body>
